+++ src/constants.ts
@@ -98,6 +98,7 @@ export const PACKAGES_TO_GET_INFO_FOR = [
   'ava',
   '@testing-library/dom',
   'cypress',
+  'turbo',
 
   'astro',
   'vue',